+++ jcl/src/java.base/share/classes/openj9/internal/criu/CRIUDumpPermission.java
@@ -1,0 +1,50 @@
+/*[INCLUDE-IF CRIU_SUPPORT]*/
+/*******************************************************************************
+ * Copyright IBM Corp. and others 2023
+ *
+ * This program and the accompanying materials are made available under
+ * the terms of the Eclipse Public License 2.0 which accompanies this
+ * distribution and is available at https://www.eclipse.org/legal/epl-2.0/
+ * or the Apache License, Version 2.0 which accompanies this distribution and
+ * is available at https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * This Source Code may also be made available under the following
+ * Secondary Licenses when the conditions for such availability set
+ * forth in the Eclipse Public License, v. 2.0 are satisfied: GNU
+ * General Public License, version 2 with the GNU Classpath
+ * Exception [1] and GNU General Public License, version 2 with the
+ * OpenJDK Assembly Exception [2].
+ *
+ * [1] https://www.gnu.org/software/classpath/license.html
+ * [2] https://openjdk.org/legal/assembly-exception.html
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0-only WITH Classpath-exception-2.0 OR GPL-2.0-only WITH OpenJDK-assembly-exception-1.0
+ *******************************************************************************/
+package openj9.internal.criu;
+
+import java.security.BasicPermission;
+
+/**
+ * The permission class for operations on the openj9.internal.criu.InternalCRIUSupport class.
+ * Allowing code access to this permission will allow applications to trigger
+ * CRIU dumps.
+ *
+ * A dump file will be a complete image of the application's memory. Code with read
+ * access to dump files produced by openj9.internal.criu.InternalCRIUSupport should be considered as having
+ * access to any information that was within the application at the time the dump
+ * was taken.
+ */
+public class CRIUDumpPermission extends BasicPermission {
+
+	/**
+	 * Serial version UID
+	 */
+	private static final long serialVersionUID = -2202468258941159543L;
+
+	/**
+	 * Creates CRIUDumpPermission.
+	 */
+	public CRIUDumpPermission() {
+		super("CRIUDumpPermission"); //$NON-NLS-1$
+	}
+}
