+++ runtime/criusupport/exports.cmake
@@ -21,9 +21,4 @@
 ################################################################################
 
 if(J9VM_OPT_CRIU_SUPPORT)
-	omr_add_exports(j9criu
-		Java_org_eclipse_openj9_criu_CRIUSupport_checkpointJVMImpl
-		Java_org_eclipse_openj9_criu_CRIUSupport_getRestoreSystemProperites;
-		Java_org_eclipse_openj9_criu_CRIUSupport_setupJNIFieldIDsAndCRIUAPI;
-	)
 endif()
