+++ runtime/criusupport/uma/criusupport_exports.xml
@@ -20,7 +20,4 @@ OpenJDK Assembly Exception [2].
 SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0-only WITH Classpath-exception-2.0 OR GPL-2.0-only WITH OpenJDK-assembly-exception-1.0
 -->
 <exports group="criusupport">
-	<export name="Java_org_eclipse_openj9_criu_CRIUSupport_checkpointJVMImpl" />
-	<export name="Java_org_eclipse_openj9_criu_CRIUSupport_getRestoreSystemProperites" />
-	<export name="Java_org_eclipse_openj9_criu_CRIUSupport_setupJNIFieldIDsAndCRIUAPI" />
 </exports>
