+++ runtime/jcl/exports.cmake
@@ -670,11 +670,14 @@ endif()
 # J9VM_OPT_CRIU_SUPPORT
 if(J9VM_OPT_CRIU_SUPPORT)
 	omr_add_exports(jclse
+		Java_openj9_internal_criu_InternalCRIUSupport_checkpointJVMImpl
 		Java_openj9_internal_criu_InternalCRIUSupport_enableCRIUSecProviderImpl
 		Java_openj9_internal_criu_InternalCRIUSupport_getCheckpointRestoreNanoTimeDeltaImpl
 		Java_openj9_internal_criu_InternalCRIUSupport_getLastRestoreTimeImpl
+		Java_openj9_internal_criu_InternalCRIUSupport_getRestoreSystemProperites;
 		Java_openj9_internal_criu_InternalCRIUSupport_isCheckpointAllowedImpl
 		Java_openj9_internal_criu_InternalCRIUSupport_isCRIUSupportEnabledImpl
+		Java_openj9_internal_criu_InternalCRIUSupport_setupJNIFieldIDsAndCRIUAPI;
 	)
 endif()
